requirements.txt
django_js_reverse
python_shout
